# lib/api-zod/src/generated/types/calculationRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnnualResult } from './annualResult';
import type { BaselineResults } from './baselineResults';
import type { InteractiveResult } from './interactiveResult';
import type { ResultSummary } from './resultSummary';
import type { SimulationUpdate } from './simulationUpdate';

export interface CalculationRun {
  interactive?: InteractiveResult;
  id: string;
  simulationId: string;
  inputVersion: number;
  engineVersion: string;
  method: string;
  createdAt: string;
  inputs: SimulationUpdate;
  annual: AnnualResult[];
  warnings: string[];
  explanations: string[];
  baseline: BaselineResults;
  summary: ResultSummary;
}