# lib/api-zod/src/generated/types/interactiveCalibration.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InteractiveCalibration {
  /** @nullable */
  baselineCollectionPercent: number | null;
  /** @nullable */
  impliedHistoricalBillings: number | null;
  /** @nullable */
  inferredAverageTariff: number | null;
  /** @nullable */
  appliedTariff: number | null;
  /** @nullable */
  baselineNrw: number | null;
  /** @nullable */
  historicalMonths: number | null;
}